Updated the way model configs are handled
