+++ src/hssm/__init__.py
@@ -3,6 +3,7 @@
 import logging
 import sys
 
+from .config import show_defaults
 from .datasets import load_data
 from .hssm import HSSM
 from .simulator import simulate_data
@@ -13,9 +14,4 @@ _logger.setLevel(logging.INFO)
 handler = logging.StreamHandler(stream=sys.stdout)
 _logger.addHandler(handler)
 
-__all__ = [
-    "HSSM",
-    "load_data",
-    "simulate_data",
-    "set_floatX",
-]
+__all__ = ["HSSM", "load_data", "simulate_data", "set_floatX", "show_defaults"]
